Refactored the name of the hpp files that contained the action message definition.
Change from bernoulli to transmission vector since most of the system has already been refactored with this name
